syntax errors
On the rules level variables were not being printed with the correct dollar binding. For the moment this is a temporary solution, I should in the future restructure all the dkPretty to take a list of dollar bindings for rules. Adding requires is still TODO and I still need to find a workaround for not having matching modulo AC